+++ src/test/java/shopping/TestConfig.java
@@ -3,7 +3,7 @@ package shopping;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import shopping.cart.component.ExchangeRateProvider;
-import shopping.cart.service.currency.MockExchangeRateProvider;
+import shopping.cart.component.MockExchangeRateProvider;
 
 @TestConfiguration
 public class TestConfig {
