+++ src/test/java/shopping/cart/component/DefaultExchangeRateProviderTest.java
@@ -1,4 +1,4 @@
-package shopping.cart.service.currency;
+package shopping.cart.component;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -10,7 +10,6 @@ import org.springframework.test.web.client.ExpectedCount;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.test.web.client.response.MockRestResponseCreators;
 import org.springframework.web.client.RestTemplate;
-import shopping.cart.component.DefaultExchangeRateProvider;
 import shopping.cart.domain.vo.ExchangeRate;
 
 import static org.assertj.core.api.Assertions.assertThat;
